requirements.txt
Flask
flask-cors
opencv-python
mediapipe
scipy
torch
torchvision
grad-cam
Pillow
matplotlib
numpy
imageio
imageio-ffmpeg
requests
gradio
